Common: Cleanup, Enforce minimum of 16 sprites

+++ Source/Sprites.hpp
@@ -292,7 +292,7 @@ struct sSprite {
     // Perhaps a linked list could replace mSprite and we use an iterator here
     union {
 		int32 field_5E_Squad;
-        int32 field_5E;			// Index inside mSprites (if 5D == false)
+        size_t field_5E;			// Index inside mSprites (if 5D == false)
         sMission_Troop *field_5E_SoldierAllocated;  // Soldier Ptr who killed sprite
     };
